Adds support for Rails 7.
Based on:
https://pagertree.com/2021/04/13/rails-7-attr-encrypted-migration/

Modifies attr_encrypted to work with Rails 7. Rails 7 includes Active
Record Encryption which has a number of name clashes with
attr_encrypted. attr_encrypted is no longer maintained so, in order to
migrate to Rails 7 we must make some modifications.

In the longer term this can be used to migrate to Active Record
Encryption.